Generate tracks layers and merge with map ones
