+++ app/src/_map/module/module.selectors.js
@@ -1,0 +1,1 @@
+export const getTemporalExtent = state => state.map.module.temporalExtent;
