+++ app/src/_map/tracks/tracks.reducer.js
@@ -7,50 +7,8 @@ import {
   REMOVE_TRACK
 } from './tracks.actions';
 
-const UPDATE_TRACKS_STYLES = 'TODO: MOVE THIS TO SELECTOR';
-
 const initialState = {
   data: []
-};
-
-const filterGeojsonByTimerange = (geojson, { start, end }) => {
-  if (!geojson || !geojson.features) return null;
-  const featuresFiltered = geojson.features.reduce((acc, feature) => {
-    const hasTimes = feature.properties.coordinateProperties.times && feature.properties.coordinateProperties.times.length > 0;
-    if (hasTimes) {
-      const filtered = feature.geometry.coordinates.reduce((acc2, coordinate, index) => {
-        const timeCoordinate = feature.properties.coordinateProperties.times[index];
-        const isInTimeline = timeCoordinate > start && timeCoordinate < end;
-        if (isInTimeline) {
-          acc2.coordinates.push(coordinate);
-          acc2.times.push(timeCoordinate);
-        }
-        return acc2;
-      }, { coordinates: [], times: [] });
-      if (!filtered.coordinates.length) return acc;
-
-      const filteredFeature = {
-        ...feature,
-        geometry: {
-          ...feature.geometry,
-          coordinates: filtered.coordinates
-        },
-        properties: {
-          ...feature.properties,
-          coordinateProperties: {
-            times: filtered.times
-          }
-        }
-      };
-      acc.push(filteredFeature);
-    }
-    return acc;
-  }, []);
-  const geojsonFiltered = {
-    ...geojson,
-    features: featuresFiltered
-  };
-  return geojsonFiltered;
 };
 
 const tracksReducer = (state = initialState, action) => {
@@ -78,46 +36,6 @@ const tracksReducer = (state = initialState, action) => {
         track.id !== removedTrackId
       );
       return { ...state, data };
-    }
-
-    case UPDATE_TRACKS_STYLES: {
-      const styles = state.data.reduce((acc, layer) => {
-        if (!layer.data) return acc;
-
-        const source = `${layer.id}Track`;
-        const style = {
-          sources: {
-            [source]: {
-              tye: 'geojson',
-              data: filterGeojsonByTimerange(layer.data, state.temporalExtent)
-            }
-          },
-          layers: [
-            {
-              source,
-              type: 'line',
-              interactive: true,
-              layout: {},
-              paint: {
-                'line-width': 1,
-                'line-color': '#000'
-              }
-            }
-          ]
-        };
-        return {
-          sources: {
-            ...acc.sources,
-            ...style.sources
-          },
-          layers: {
-            ...acc.layers,
-            ...style.layers
-          }
-        };
-      }, { sources: [], layers: [] });
-      console.log(styles)
-      return { ...state, styles };
     }
 
     default:
