+++ package.json
@@ -43,7 +43,7 @@
     "d3": "^4.10.0",
     "d3-ease": "^1.0.3",
     "dotenv": "^4.0.0",
-    "immutable": "^3.8.2",
+    "immutable": "4.0.0-rc.12",
     "iso-3166-1-alpha-2": "^1.0.0",
     "lodash": "^4.17.4",
     "mapbox-gl": "^0.53.0",
